ml_experiments/resnet18_finetune/requirements.txt
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
scikit-learn==1.4.2
pillow==10.3.0
matplotlib==3.8.4
seaborn==0.13.2
mlflow==2.12.1
pyyaml==6.0.1
tqdm==4.66.4
pandas==2.2.2
